VCV-175: address comments
